fix UI for health and speed of player

+++ Assets/Scripts/Interfaces/IHealthUp.cs
@@ -1,4 +1,4 @@
 public interface IHealthUp
 {
-    public float HealthUp();
+    public int HealthUp();
 }
